backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pydantic
python-multipart
requests
python-dotenv
